refactor(weapons): cut over weapon runtime from gameplay
