+++ src/crimson/gameplay.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 import math
-from typing import TYPE_CHECKING, Callable, Protocol, Sequence
+from typing import TYPE_CHECKING, Sequence
 
-from grim.color import RGBA
 from grim.geom import Vec2
 from grim.rand import Crand
 from .bonuses.freeze import DeferredFreezeCorpseFx
@@ -23,16 +22,22 @@ from .projectiles import (
     ProjectilePool,
     ProjectileTypeId,
     SecondaryProjectilePool,
-    SecondaryProjectileTypeId,
 )
-from .weapons import (
-    WEAPON_BY_ID,
-    WEAPON_TABLE,
-    Weapon,
-    WeaponId,
-    projectile_type_id_from_weapon_id,
-    weapon_entry_for_projectile_type_id,
+from .weapon_runtime import (
+    most_used_weapon_id_for_player,
+    owner_id_for_player as _owner_id_for_player,
+    owner_id_for_player_projectiles as _owner_id_for_player_projectiles,
+    player_fire_weapon,
+    player_start_reload,
+    player_swap_alt_weapon,
+    projectile_spawn as _projectile_spawn,
+    spawn_projectile_ring as _spawn_projectile_ring,
+    weapon_entry as _weapon_entry,
+    weapon_assign_player,
+    weapon_pick_random_available,
+    weapon_refresh_available,
 )
+from .weapons import WEAPON_TABLE, WeaponId
 from .sim.input import PlayerInput
 from .sim.state_types import PERK_COUNT_SIZE, PlayerState
 
@@ -40,11 +45,14 @@ if TYPE_CHECKING:
     from .persistence.save_status import GameStatus
 
 
-class _HasPos(Protocol):
-    pos: Vec2
-
-
 WEAPON_COUNT_SIZE = max(int(entry.weapon_id) for entry in WEAPON_TABLE) + 1
+# Compatibility shim: keep weapon runtime API available from `crimson.gameplay`
+# while call sites are migrated.
+_COMPAT_WEAPON_RUNTIME_EXPORTS = (
+    most_used_weapon_id_for_player,
+    weapon_pick_random_available,
+    weapon_refresh_available,
+)
 
 
 @dataclass(slots=True)
@@ -56,7 +64,6 @@ class BonusTimers:
     freeze: float = 0.0
 
 
-WEAPON_DROP_ID_COUNT = 0x21  # weapon ids 1..33
 _RELOAD_PRELOAD_UNDERFLOW_EPS = 1e-7
 
 
@@ -277,711 +284,6 @@ def survival_enforce_reward_weapon_guard(state: GameplayState, players: Sequence
             weapon_assign_player(player, int(WeaponId.PISTOL))
         if weapon_id == int(WeaponId.SHRINKIFIER_5K) and guard_id != int(WeaponId.SHRINKIFIER_5K):
             weapon_assign_player(player, int(WeaponId.PISTOL))
-
-
-def _owner_id_for_player(player_index: int) -> int:
-    # crimsonland.exe uses -1/-2/-3 for players (and sometimes -100 in demo paths).
-    return -1 - int(player_index)
-
-
-def _owner_id_for_player_projectiles(state: "GameplayState", player_index: int) -> int:
-    if not state.friendly_fire_enabled:
-        return -100
-    return _owner_id_for_player(player_index)
-
-
-def _weapon_entry(weapon_id: int) -> Weapon | None:
-    return WEAPON_BY_ID.get(int(weapon_id))
-
-
-def weapon_refresh_available(state: "GameplayState") -> None:
-    """Rebuild `weapon_table[weapon_id].unlocked` equivalents from quest progression.
-
-    Port of `weapon_refresh_available` (0x00452e40).
-    """
-
-    unlock_index = 0
-    unlock_index_full = 0
-    status = state.status
-    if status is not None:
-        try:
-            unlock_index = int(status.quest_unlock_index)
-        except Exception:
-            unlock_index = 0
-        try:
-            unlock_index_full = int(status.quest_unlock_index_full)
-        except Exception:
-            unlock_index_full = 0
-
-    game_mode = int(state.game_mode)
-    if (
-        int(state._weapon_available_game_mode) == game_mode
-        and int(state._weapon_available_unlock_index) == unlock_index
-        and int(state._weapon_available_unlock_index_full) == unlock_index_full
-    ):
-        return
-
-    # Clear unlocked flags.
-    available = state.weapon_available
-    for idx in range(len(available)):
-        available[idx] = False
-
-    # Pistol is always available.
-    pistol_id = int(WeaponId.PISTOL)
-    if 0 <= pistol_id < len(available):
-        available[pistol_id] = True
-
-    # Unlock weapons from the quest list (first `quest_unlock_index` entries).
-    if unlock_index > 0:
-        try:
-            from .quests import all_quests
-
-            quests = all_quests()
-        except Exception:
-            quests = []
-
-        for quest in quests[:unlock_index]:
-            weapon_id = int(getattr(quest, "unlock_weapon_id", 0) or 0)
-            if 0 < weapon_id < len(available):
-                available[weapon_id] = True
-
-    # Survival default loadout: Assault Rifle, Shotgun, Submachine Gun.
-    if game_mode == int(GameMode.SURVIVAL):
-        for weapon_id in (WeaponId.ASSAULT_RIFLE, WeaponId.SHOTGUN, WeaponId.SUBMACHINE_GUN):
-            idx = int(weapon_id)
-            if 0 <= idx < len(available):
-                available[idx] = True
-
-    # Secret unlock: Splitter Gun (weapon id 29) becomes available once the hardcore
-    # unlock track reaches stage 5 (quest_unlock_index_full >= 40).
-    if (not state.demo_mode_active) and unlock_index_full >= 0x28:
-        splitter_id = int(WeaponId.SPLITTER_GUN)
-        if 0 <= splitter_id < len(available):
-            available[splitter_id] = True
-
-    state._weapon_available_game_mode = game_mode
-    state._weapon_available_unlock_index = unlock_index
-    state._weapon_available_unlock_index_full = unlock_index_full
-
-
-def weapon_pick_random_available(state: "GameplayState") -> int:
-    """Select a random available weapon id (1..33).
-
-    Port of `weapon_pick_random_available` (0x00452cd0).
-    """
-
-    weapon_refresh_available(state)
-    status = state.status
-
-    for _ in range(1000):
-        base_rand = int(state.rng.rand())
-        weapon_id = base_rand % WEAPON_DROP_ID_COUNT + 1
-
-        # Bias: used weapons have a 50% chance to reroll once.
-        if status is not None:
-            try:
-                if status.weapon_usage_count(weapon_id) != 0:
-                    if (int(state.rng.rand()) & 1) == 0:
-                        base_rand = int(state.rng.rand())
-                        weapon_id = base_rand % WEAPON_DROP_ID_COUNT + 1
-            except Exception:
-                pass
-
-        if not (0 <= weapon_id < len(state.weapon_available)):
-            continue
-        if not state.weapon_available[weapon_id]:
-            continue
-
-        # Quest 5-10 special-case: suppress Ion Cannon.
-        if (
-            int(state.game_mode) == int(GameMode.QUESTS)
-            and int(state.quest_stage_major) == 5
-            and int(state.quest_stage_minor) == 10
-            and weapon_id == int(WeaponId.ION_CANNON)
-        ):
-            continue
-
-        return weapon_id
-
-    return int(WeaponId.PISTOL)
-
-
-def _projectile_meta_for_type_id(type_id: int) -> float:
-    entry = weapon_entry_for_projectile_type_id(int(type_id))
-    meta = entry.projectile_meta if entry is not None else None
-    return float(meta if meta is not None else 45.0)
-
-
-@dataclass(slots=True)
-class _WeaponAssignCtx:
-    player: PlayerState
-    clip_size: int
-
-
-_WeaponAssignClipModifier = Callable[[_WeaponAssignCtx], None]
-
-
-def _weapon_assign_clip_ammo_maniac(ctx: _WeaponAssignCtx) -> None:
-    if perk_active(ctx.player, PerkId.AMMO_MANIAC):
-        ctx.clip_size += max(1, int(float(ctx.clip_size) * 0.25))
-
-
-def _weapon_assign_clip_my_favourite_weapon(ctx: _WeaponAssignCtx) -> None:
-    if perk_active(ctx.player, PerkId.MY_FAVOURITE_WEAPON):
-        ctx.clip_size += 2
-
-
-_WEAPON_ASSIGN_CLIP_MODIFIERS: tuple[_WeaponAssignClipModifier, ...] = (
-    _weapon_assign_clip_ammo_maniac,
-    _weapon_assign_clip_my_favourite_weapon,
-)
-
-
-def weapon_assign_player(player: PlayerState, weapon_id: int, *, state: GameplayState | None = None) -> None:
-    """Assign weapon and reset per-weapon runtime state (ammo/cooldowns)."""
-
-    weapon_id = int(weapon_id)
-    if state is not None and state.status is not None and not state.demo_mode_active:
-        try:
-            state.status.increment_weapon_usage(weapon_id)
-        except Exception:
-            pass
-
-    weapon = _weapon_entry(weapon_id)
-    player.weapon_id = weapon_id
-
-    clip_size = int(weapon.clip_size) if weapon is not None and weapon.clip_size is not None else 0
-    clip_ctx = _WeaponAssignCtx(player=player, clip_size=max(0, clip_size))
-    for modifier in _WEAPON_ASSIGN_CLIP_MODIFIERS:
-        modifier(clip_ctx)
-    player.clip_size = max(0, int(clip_ctx.clip_size))
-    player.ammo = float(player.clip_size)
-    player.weapon_reset_latch = 0
-    player.reload_active = False
-    player.reload_timer = 0.0
-    player.reload_timer_max = 0.0
-    player.shot_cooldown = 0.0
-    player.aux_timer = 2.0
-
-    if state is not None and weapon is not None:
-        from .weapon_sfx import resolve_weapon_sfx_ref
-
-        key = resolve_weapon_sfx_ref(weapon.reload_sound)
-        if key is not None:
-            state.sfx_queue.append(key)
-
-
-def most_used_weapon_id_for_player(state: GameplayState, *, player_index: int, fallback_weapon_id: int) -> int:
-    """Return a 1-based weapon id for the player's most-used weapon."""
-
-    idx = int(player_index)
-    if 0 <= idx < len(state.weapon_shots_fired):
-        counts = state.weapon_shots_fired[idx]
-        if counts:
-            start = 1 if len(counts) > 1 else 0
-            best = max(range(start, len(counts)), key=lambda i: int(counts[i]))
-            if int(counts[best]) > 0:
-                return int(best)
-    return int(fallback_weapon_id)
-
-
-def player_swap_alt_weapon(player: PlayerState) -> bool:
-    """Swap primary and alternate weapon runtime blocks (Alternate Weapon perk)."""
-
-    if player.alt_weapon_id is None:
-        return False
-    (
-        player.weapon_id,
-        player.clip_size,
-        player.reload_active,
-        player.ammo,
-        player.reload_timer,
-        player.shot_cooldown,
-        player.reload_timer_max,
-        player.alt_weapon_id,
-        player.alt_clip_size,
-        player.alt_reload_active,
-        player.alt_ammo,
-        player.alt_reload_timer,
-        player.alt_shot_cooldown,
-        player.alt_reload_timer_max,
-    ) = (
-        player.alt_weapon_id,
-        player.alt_clip_size,
-        player.alt_reload_active,
-        player.alt_ammo,
-        player.alt_reload_timer,
-        player.alt_shot_cooldown,
-        player.alt_reload_timer_max,
-        player.weapon_id,
-        player.clip_size,
-        player.reload_active,
-        player.ammo,
-        player.reload_timer,
-        player.shot_cooldown,
-        player.reload_timer_max,
-    )
-    return True
-
-
-def player_start_reload(player: PlayerState, state: GameplayState) -> None:
-    """Start or refresh a reload timer (`player_start_reload` @ 0x00413430)."""
-
-    if player.reload_active and (
-        perk_active(player, PerkId.AMMUNITION_WITHIN) or perk_active(player, PerkId.REGRESSION_BULLETS)
-    ):
-        return
-
-    weapon = _weapon_entry(player.weapon_id)
-    reload_time = float(weapon.reload_time) if weapon is not None and weapon.reload_time is not None else 0.0
-
-    if not player.reload_active:
-        player.reload_active = True
-
-    if perk_active(player, PerkId.FASTLOADER):
-        reload_time *= 0.7
-    if state.bonuses.weapon_power_up > 0.0:
-        reload_time *= 0.6
-
-    player.reload_timer = max(0.0, reload_time)
-    player.reload_timer_max = player.reload_timer
-
-
-def _spawn_projectile_ring(
-    state: GameplayState,
-    origin: _HasPos,
-    *,
-    count: int,
-    angle_offset: float,
-    type_id: int,
-    owner_id: int,
-    players: list[PlayerState] | None = None,
-) -> None:
-    if count <= 0:
-        return
-    step = math.tau / float(count)
-    for idx in range(count):
-        _projectile_spawn(
-            state,
-            players=players,
-            pos=origin.pos,
-            angle=float(idx) * step + float(angle_offset),
-            type_id=int(type_id),
-            owner_id=int(owner_id),
-        )
-
-
-def _fire_bullets_active(players: list[PlayerState] | None) -> bool:
-    # Native `projectile_spawn` checks `player_state_table.fire_bullets_timer` and `player2_fire_bullets_timer`
-    # (i.e. the first two players).
-    if not players:
-        return False
-    for player in players[:2]:
-        if float(player.fire_bullets_timer) > 0.0:
-            return True
-    return False
-
-
-def _projectile_spawn(
-    state: GameplayState,
-    *,
-    players: list[PlayerState] | None,
-    pos: Vec2,
-    angle: float,
-    type_id: int,
-    owner_id: int,
-    hits_players: bool = False,
-) -> int:
-    # Mirror `projectile_spawn` (0x00420440) Fire Bullets override.
-    type_id = int(type_id)
-    owner_id = int(owner_id)
-    if (
-        (not state.bonus_spawn_guard)
-        and owner_id in (-100, -1, -2, -3)
-        and type_id != int(ProjectileTypeId.FIRE_BULLETS)
-        and _fire_bullets_active(players)
-    ):
-        type_id = int(ProjectileTypeId.FIRE_BULLETS)
-
-    meta = _projectile_meta_for_type_id(type_id)
-    return state.projectiles.spawn(
-        pos=pos,
-        angle=float(angle),
-        type_id=int(type_id),
-        owner_id=int(owner_id),
-        base_damage=float(meta),
-        hits_players=bool(hits_players),
-    )
-
-
-_NATIVE_FIRE_MUZZLE_SPRITES: dict[int, tuple[tuple[float, float, float], ...]] = {
-    int(WeaponId.PISTOL): ((25.0, 1.0, 0.23), (15.0, 2.0, 0.213)),
-    int(WeaponId.ASSAULT_RIFLE): ((25.0, 1.0, 0.23), (15.0, 2.0, 0.213)),
-    int(WeaponId.SHOTGUN): ((25.0, 1.0, 0.25), (15.0, 2.0, 0.223)),
-    int(WeaponId.SAWED_OFF_SHOTGUN): ((25.0, 1.0, 0.26), (15.0, 2.0, 0.233)),
-    int(WeaponId.SUBMACHINE_GUN): ((25.0, 1.0, 0.23), (15.0, 2.0, 0.213)),
-    int(WeaponId.GAUSS_GUN): ((25.0, 1.0, 0.33), (15.0, 2.0, 0.263)),
-    int(WeaponId.ROCKET_LAUNCHER): ((25.0, 1.0, 0.34), (15.0, 2.0, 0.283)),
-    int(WeaponId.SEEKER_ROCKETS): ((25.0, 1.0, 0.31), (15.0, 2.0, 0.243)),
-    int(WeaponId.MINI_ROCKET_SWARMERS): ((25.0, 1.0, 0.34), (15.0, 2.0, 0.283)),
-    int(WeaponId.ROCKET_MINIGUN): ((25.0, 1.0, 0.34),),
-    int(WeaponId.JACKHAMMER): ((15.0, 2.0, 0.223),),
-    int(WeaponId.SHRINKIFIER_5K): ((25.0, 1.0, 0.23), (15.0, 2.0, 0.213)),
-    int(WeaponId.GAUSS_SHOTGUN): ((25.0, 1.0, 0.33), (15.0, 2.0, 0.263)),
-}
-
-_NATIVE_FIRE_MUZZLE_AFTER_PROJECTILE: frozenset[int] = frozenset(
-    {
-        int(WeaponId.PISTOL),
-        int(WeaponId.SHRINKIFIER_5K),
-    }
-)
-
-
-def _spawn_native_fire_muzzle_sprites(
-    *,
-    state: GameplayState,
-    weapon_id: int,
-    muzzle: Vec2,
-    aim_heading: float,
-    fire_bullets_active: bool,
-) -> None:
-    if fire_bullets_active:
-        specs: tuple[tuple[float, float, float], ...] = ((25.0, 1.0, 0.413),)
-    else:
-        specs = _NATIVE_FIRE_MUZZLE_SPRITES.get(int(weapon_id), ())
-    if not specs:
-        return
-
-    for speed, scale, alpha in specs:
-        state.sprite_effects.spawn(
-            pos=muzzle,
-            vel=Vec2.from_heading(aim_heading) * float(speed),
-            scale=float(scale),
-            color=RGBA(0.5, 0.5, 0.5, float(alpha)),
-        )
-
-
-def player_fire_weapon(
-    player: PlayerState,
-    input_state: PlayerInput,
-    dt: float,
-    state: GameplayState,
-    *,
-    detail_preset: int = 5,
-    creatures: Sequence[Damageable] | None = None,
-) -> None:
-    dt = float(dt)
-
-    weapon_id = int(player.weapon_id)
-    weapon = _weapon_entry(weapon_id)
-    if weapon is None:
-        return
-
-    if player.shot_cooldown > 0.0:
-        return
-    if not input_state.fire_down:
-        return
-
-    ammo_cost = 1.0
-    is_fire_bullets = float(player.fire_bullets_timer) > 0.0
-    if player.reload_timer > 0.0:
-        if player.experience <= 0:
-            return
-        if perk_active(player, PerkId.REGRESSION_BULLETS):
-            ammo_class = int(weapon.ammo_class) if weapon.ammo_class is not None else 0
-
-            reload_time = float(weapon.reload_time) if weapon.reload_time is not None else 0.0
-            factor = 4.0 if ammo_class == 1 else 200.0
-            player.experience = int(float(player.experience) - reload_time * factor)
-            if player.experience < 0:
-                player.experience = 0
-        elif perk_active(player, PerkId.AMMUNITION_WITHIN):
-            ammo_class = int(weapon.ammo_class) if weapon.ammo_class is not None else 0
-
-            from .player_damage import player_take_damage
-
-            cost = 0.15 if ammo_class == 1 else 1.0
-            player_take_damage(state, player, cost, dt=dt, rand=state.rng.rand)
-        else:
-            return
-
-    pellet_count = int(weapon.pellet_count) if weapon.pellet_count is not None else 0
-    fire_bullets_weapon = weapon_entry_for_projectile_type_id(int(ProjectileTypeId.FIRE_BULLETS))
-
-    shot_cooldown = float(weapon.shot_cooldown) if weapon.shot_cooldown is not None else 0.0
-    weapon_spread_heat = float(weapon.spread_heat_inc) if weapon.spread_heat_inc is not None else 0.0
-    fire_bullets_spread_heat = weapon_spread_heat
-    if fire_bullets_weapon is not None and fire_bullets_weapon.spread_heat_inc is not None:
-        fire_bullets_spread_heat = float(fire_bullets_weapon.spread_heat_inc)
-
-    if is_fire_bullets and pellet_count == 1 and fire_bullets_weapon is not None:
-        shot_cooldown = (
-            float(fire_bullets_weapon.shot_cooldown) if fire_bullets_weapon.shot_cooldown is not None else 0.0
-        )
-
-    spread_heat_base = fire_bullets_spread_heat if is_fire_bullets else weapon_spread_heat
-    spread_inc = spread_heat_base * 1.3
-
-    if perk_active(player, PerkId.FASTSHOT):
-        shot_cooldown *= 0.88
-    if perk_active(player, PerkId.SHARPSHOOTER):
-        shot_cooldown *= 1.05
-    player.shot_cooldown = max(0.0, shot_cooldown)
-
-    aim = input_state.aim
-    aim_delta = aim - player.pos
-    aim_heading = aim_delta.to_heading()
-
-    muzzle = player.pos + Vec2.from_heading(aim_heading).rotated(-0.150915) * 16.0
-    state.effects.spawn_shell_casing(
-        pos=muzzle,
-        aim_heading=aim_heading,
-        weapon_flags=int(weapon.flags or 0),
-        rand=state.rng.rand,
-        detail_preset=int(detail_preset),
-    )
-
-    dist = aim_delta.length()
-    max_offset = dist * float(player.spread_heat) * 0.5
-    dir_angle = float(int(state.rng.rand()) & 0x1FF) * (math.tau / 512.0)
-    mag = float(int(state.rng.rand()) & 0x1FF) * (1.0 / 512.0)
-    offset = max_offset * mag
-    aim_jitter = aim + Vec2.from_angle(dir_angle) * offset
-    shot_angle = (aim_jitter - player.pos).to_heading()
-    particle_angle = Vec2.from_heading(shot_angle).to_angle()
-    if weapon_id in (WeaponId.FLAMETHROWER, WeaponId.BLOW_TORCH, WeaponId.HR_FLAMER):
-        particle_angle = Vec2.from_heading(aim_heading).to_angle()
-
-    # Native `player_fire_weapon` consumes one RNG draw for shot SFX variant
-    # selection on every non-Fire-Bullets shot.
-    if not is_fire_bullets:
-        state.rng.rand()
-
-    owner_id = _owner_id_for_player(player.index)
-    projectile_owner_id = _owner_id_for_player_projectiles(state, player.index)
-    shot_count = 1
-    spawn_muzzle_after_projectile = bool(is_fire_bullets) or int(weapon_id) in _NATIVE_FIRE_MUZZLE_AFTER_PROJECTILE
-    if not spawn_muzzle_after_projectile:
-        _spawn_native_fire_muzzle_sprites(
-            state=state,
-            weapon_id=int(weapon_id),
-            muzzle=muzzle,
-            aim_heading=float(aim_heading),
-            fire_bullets_active=bool(is_fire_bullets),
-        )
-
-    # `player_fire_weapon` (crimsonland.exe) uses weapon-specific extra angular jitter for pellet
-    # weapons. This is separate from aim-point jitter driven by `player.spread_heat`.
-    def _pellet_jitter_step(weapon_id: int) -> float:
-        weapon_id = int(weapon_id)
-        if weapon_id == WeaponId.SHOTGUN:
-            return 0.0013
-        if weapon_id == WeaponId.SAWED_OFF_SHOTGUN:
-            return 0.004
-        if weapon_id == WeaponId.JACKHAMMER:
-            return 0.0013
-        return 0.0015
-
-    if is_fire_bullets:
-        pellets = max(0, int(pellet_count))
-        shot_count = pellets
-        meta = _projectile_meta_for_type_id(ProjectileTypeId.FIRE_BULLETS)
-        for _ in range(pellets):
-            angle = shot_angle + float(int(state.rng.rand()) % 200 - 100) * 0.0015
-            state.projectiles.spawn(
-                pos=muzzle,
-                angle=angle,
-                type_id=ProjectileTypeId.FIRE_BULLETS,
-                owner_id=projectile_owner_id,
-                base_damage=meta,
-            )
-    elif weapon_id == WeaponId.ROCKET_LAUNCHER:
-        # Rocket Launcher -> secondary type 1.
-        state.secondary_projectiles.spawn(
-            pos=muzzle,
-            angle=shot_angle,
-            type_id=SecondaryProjectileTypeId.ROCKET,
-            owner_id=owner_id,
-        )
-    elif weapon_id == WeaponId.SEEKER_ROCKETS:
-        # Seeker Rockets -> secondary type 2.
-        state.secondary_projectiles.spawn(
-            pos=muzzle,
-            angle=shot_angle,
-            type_id=SecondaryProjectileTypeId.HOMING_ROCKET,
-            owner_id=owner_id,
-            target_hint=aim,
-            creatures=creatures,
-        )
-    elif weapon_id == WeaponId.MINI_ROCKET_SWARMERS:
-        # Mini-Rocket Swarmers -> secondary type 2 (fires the full clip in a spread).
-        rocket_count = max(1, int(player.ammo))
-        step = float(rocket_count) * (math.pi / 3.0)
-        angle = (shot_angle - math.pi) - step * float(rocket_count) * 0.5
-        for _ in range(rocket_count):
-            state.secondary_projectiles.spawn(
-                pos=muzzle,
-                angle=angle,
-                type_id=SecondaryProjectileTypeId.HOMING_ROCKET,
-                owner_id=owner_id,
-                target_hint=aim,
-                creatures=creatures,
-            )
-            angle += step
-        ammo_cost = float(rocket_count)
-        shot_count = rocket_count
-    elif weapon_id == WeaponId.ROCKET_MINIGUN:
-        # Rocket Minigun -> secondary type 4.
-        state.secondary_projectiles.spawn(
-            pos=muzzle,
-            angle=shot_angle,
-            type_id=SecondaryProjectileTypeId.ROCKET_MINIGUN,
-            owner_id=owner_id,
-        )
-    elif weapon_id == WeaponId.FLAMETHROWER:
-        # Flamethrower -> fast particle weapon (style 0), fractional ammo drain.
-        state.particles.spawn_particle(pos=muzzle, angle=particle_angle, intensity=1.0, owner_id=owner_id)
-        ammo_cost = 0.1
-    elif weapon_id == WeaponId.BLOW_TORCH:
-        # Blow Torch -> fast particle weapon (style 1), fractional ammo drain.
-        particle_id = state.particles.spawn_particle(pos=muzzle, angle=particle_angle, intensity=1.0, owner_id=owner_id)
-        state.particles.entries[particle_id].style_id = 1
-        ammo_cost = 0.05
-    elif weapon_id == WeaponId.HR_FLAMER:
-        # HR Flamer -> fast particle weapon (style 2), fractional ammo drain.
-        particle_id = state.particles.spawn_particle(pos=muzzle, angle=particle_angle, intensity=1.0, owner_id=owner_id)
-        state.particles.entries[particle_id].style_id = 2
-        ammo_cost = 0.1
-    elif weapon_id == WeaponId.BUBBLEGUN:
-        # Bubblegun -> slow particle weapon (style 8), fractional ammo drain.
-        state.particles.spawn_particle_slow(
-            pos=muzzle,
-            angle=Vec2.from_heading(shot_angle).to_angle(),
-            owner_id=owner_id,
-        )
-        ammo_cost = 0.15
-    elif weapon_id == WeaponId.MULTI_PLASMA:
-        # Multi-Plasma: 5-shot fixed spread using type 0x09 and 0x0B.
-        # (`player_update` weapon_id==0x0a in crimsonland.exe)
-        shot_count = 5
-        # Native literals: 0.31415927 (~ pi/10), 0.5235988 (~ pi/6).
-        spread_small = math.pi / 10
-        spread_large = math.pi / 6
-        patterns: tuple[tuple[float, ProjectileTypeId], ...] = (
-            (-spread_small, ProjectileTypeId.PLASMA_RIFLE),
-            (-spread_large, ProjectileTypeId.PLASMA_MINIGUN),
-            (0.0, ProjectileTypeId.PLASMA_RIFLE),
-            (spread_large, ProjectileTypeId.PLASMA_MINIGUN),
-            (spread_small, ProjectileTypeId.PLASMA_RIFLE),
-        )
-        for angle_offset, type_id in patterns:
-            state.projectiles.spawn(
-                pos=muzzle,
-                angle=shot_angle + angle_offset,
-                type_id=type_id,
-                owner_id=projectile_owner_id,
-                base_damage=_projectile_meta_for_type_id(type_id),
-            )
-    elif weapon_id == WeaponId.PLASMA_SHOTGUN:
-        # Plasma Shotgun: 14 plasma-minigun pellets with wide jitter and random speed_scale.
-        # (`player_update` weapon_id==0x0e in crimsonland.exe)
-        shot_count = 14
-        meta = _projectile_meta_for_type_id(int(ProjectileTypeId.PLASMA_MINIGUN))
-        for _ in range(14):
-            jitter = float((int(state.rng.rand()) & 0xFF) - 0x80) * 0.002
-            proj_id = state.projectiles.spawn(
-                pos=muzzle,
-                angle=shot_angle + jitter,
-                type_id=ProjectileTypeId.PLASMA_MINIGUN,
-                owner_id=projectile_owner_id,
-                base_damage=meta,
-            )
-            state.projectiles.entries[int(proj_id)].speed_scale = 1.0 + float(int(state.rng.rand()) % 100) * 0.01
-    elif weapon_id == WeaponId.GAUSS_SHOTGUN:
-        # Gauss Shotgun: 6 gauss pellets, jitter 0.002 and speed_scale 1.4..(1.4 + 0.79).
-        # (`player_update` weapon_id==0x1e in crimsonland.exe)
-        shot_count = 6
-        meta = _projectile_meta_for_type_id(int(ProjectileTypeId.GAUSS_GUN))
-        for _ in range(6):
-            jitter = float(int(state.rng.rand()) % 200 - 100) * 0.002
-            proj_id = state.projectiles.spawn(
-                pos=muzzle,
-                angle=shot_angle + jitter,
-                type_id=ProjectileTypeId.GAUSS_GUN,
-                owner_id=projectile_owner_id,
-                base_damage=meta,
-            )
-            state.projectiles.entries[int(proj_id)].speed_scale = 1.4 + float(int(state.rng.rand()) % 0x50) * 0.01
-    elif weapon_id == WeaponId.ION_SHOTGUN:
-        # Ion Shotgun: 8 ion-minigun pellets, jitter 0.0026 and speed_scale 1.4..(1.4 + 0.79).
-        # (`player_update` weapon_id==0x1f in crimsonland.exe)
-        shot_count = 8
-        meta = _projectile_meta_for_type_id(int(ProjectileTypeId.ION_MINIGUN))
-        for _ in range(8):
-            jitter = float(int(state.rng.rand()) % 200 - 100) * 0.0026
-            proj_id = state.projectiles.spawn(
-                pos=muzzle,
-                angle=shot_angle + jitter,
-                type_id=ProjectileTypeId.ION_MINIGUN,
-                owner_id=projectile_owner_id,
-                base_damage=meta,
-            )
-            state.projectiles.entries[int(proj_id)].speed_scale = 1.4 + float(int(state.rng.rand()) % 0x50) * 0.01
-    else:
-        pellets = max(1, int(pellet_count))
-        shot_count = pellets
-        type_id = projectile_type_id_from_weapon_id(weapon_id)
-        if type_id is None:
-            return
-        meta = _projectile_meta_for_type_id(type_id)
-        jitter_step = _pellet_jitter_step(weapon_id)
-        for _ in range(pellets):
-            angle = shot_angle
-            if pellets > 1:
-                angle += float(int(state.rng.rand()) % 200 - 100) * jitter_step
-            proj_id = state.projectiles.spawn(
-                pos=muzzle,
-                angle=angle,
-                type_id=type_id,
-                owner_id=projectile_owner_id,
-                base_damage=meta,
-            )
-            # Shotgun variants randomize speed_scale per pellet (rand%100 * 0.01 + 1.0).
-            if pellets > 1 and weapon_id in (WeaponId.SHOTGUN, WeaponId.SAWED_OFF_SHOTGUN, WeaponId.JACKHAMMER):
-                state.projectiles.entries[int(proj_id)].speed_scale = 1.0 + float(int(state.rng.rand()) % 100) * 0.01
-
-    if 0 <= int(player.index) < len(state.shots_fired):
-        state.shots_fired[int(player.index)] += int(shot_count)
-        if 0 <= weapon_id < WEAPON_COUNT_SIZE:
-            state.weapon_shots_fired[int(player.index)][weapon_id] += int(shot_count)
-
-    if spawn_muzzle_after_projectile:
-        _spawn_native_fire_muzzle_sprites(
-            state=state,
-            weapon_id=int(weapon_id),
-            muzzle=muzzle,
-            aim_heading=float(aim_heading),
-            fire_bullets_active=bool(is_fire_bullets),
-        )
-
-    if not perk_active(player, PerkId.SHARPSHOOTER):
-        player.spread_heat = min(0.48, max(0.0, player.spread_heat + spread_inc))
-
-    muzzle_inc = weapon_spread_heat
-    if is_fire_bullets and pellet_count == 1:
-        muzzle_inc = fire_bullets_spread_heat
-    player.muzzle_flash_alpha = min(1.0, player.muzzle_flash_alpha)
-    player.muzzle_flash_alpha = min(1.0, player.muzzle_flash_alpha + muzzle_inc)
-    player.muzzle_flash_alpha = min(0.8, player.muzzle_flash_alpha)
-
-    player.shot_seq += 1
-    if state.bonuses.reflex_boost <= 0.0 and not is_fire_bullets:
-        # Native allows ammo to cross below zero for reload-time firing paths
-        # (for example Regression Bullets), and replay checkpoints rely on that.
-        player.ammo = float(player.ammo) - float(ammo_cost)
-    if player.ammo <= 0.0 and player.reload_timer <= 0.0:
-        player_start_reload(player, state)
 
 
 def player_update(
