+++ src/crimson/perks/impl/ammo_maniac.py
@@ -3,11 +3,10 @@ from __future__ import annotations
 from ..runtime.apply_context import PerkApplyCtx
 from ..runtime.hook_types import PerkHooks
 from ..ids import PerkId
+from ...weapon_runtime.assign import weapon_assign_player
 
 
 def apply_ammo_maniac(ctx: PerkApplyCtx) -> None:
-    from ...gameplay import weapon_assign_player
-
     if len(ctx.players) > 1:
         for player in ctx.players[1:]:
             player.perk_counts[:] = ctx.owner.perk_counts
