+++ src/crimson/perks/impl/random_weapon.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from ...weapon_runtime.assign import weapon_assign_player
+from ...weapon_runtime.availability import weapon_pick_random_available
 from ...weapons import WeaponId
 from ..runtime.apply_context import PerkApplyCtx
 from ..runtime.hook_types import PerkHooks
@@ -7,8 +9,6 @@ from ..ids import PerkId
 
 
 def apply_random_weapon(ctx: PerkApplyCtx) -> None:
-    from ...gameplay import weapon_assign_player, weapon_pick_random_available
-
     current = int(ctx.owner.weapon_id)
     weapon_id = int(current)
     for _ in range(100):
