+++ src/crimson/weapon_runtime/assign.py
@@ -1,0 +1,148 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Callable
+
+from ..perks import PerkId
+from ..perks.helpers import perk_active
+from ..sim.state_types import GameplayState, PlayerState
+from ..weapons import WEAPON_BY_ID, Weapon
+
+
+def weapon_entry(weapon_id: int) -> Weapon | None:
+    return WEAPON_BY_ID.get(int(weapon_id))
+
+
+@dataclass(slots=True)
+class _WeaponAssignCtx:
+    player: PlayerState
+    clip_size: int
+
+
+_WeaponAssignClipModifier = Callable[[_WeaponAssignCtx], None]
+
+
+def _weapon_assign_clip_ammo_maniac(ctx: _WeaponAssignCtx) -> None:
+    if perk_active(ctx.player, PerkId.AMMO_MANIAC):
+        ctx.clip_size += max(1, int(float(ctx.clip_size) * 0.25))
+
+
+def _weapon_assign_clip_my_favourite_weapon(ctx: _WeaponAssignCtx) -> None:
+    if perk_active(ctx.player, PerkId.MY_FAVOURITE_WEAPON):
+        ctx.clip_size += 2
+
+
+_WEAPON_ASSIGN_CLIP_MODIFIERS: tuple[_WeaponAssignClipModifier, ...] = (
+    _weapon_assign_clip_ammo_maniac,
+    _weapon_assign_clip_my_favourite_weapon,
+)
+
+
+def weapon_assign_player(player: PlayerState, weapon_id: int, *, state: GameplayState | None = None) -> None:
+    """Assign weapon and reset per-weapon runtime state (ammo/cooldowns)."""
+
+    weapon_id = int(weapon_id)
+    if state is not None and state.status is not None and not state.demo_mode_active:
+        try:
+            state.status.increment_weapon_usage(weapon_id)
+        except Exception:
+            pass
+
+    weapon = weapon_entry(weapon_id)
+    player.weapon_id = weapon_id
+
+    clip_size = int(weapon.clip_size) if weapon is not None and weapon.clip_size is not None else 0
+    clip_ctx = _WeaponAssignCtx(player=player, clip_size=max(0, clip_size))
+    for modifier in _WEAPON_ASSIGN_CLIP_MODIFIERS:
+        modifier(clip_ctx)
+    player.clip_size = max(0, int(clip_ctx.clip_size))
+    player.ammo = float(player.clip_size)
+    player.weapon_reset_latch = 0
+    player.reload_active = False
+    player.reload_timer = 0.0
+    player.reload_timer_max = 0.0
+    player.shot_cooldown = 0.0
+    player.aux_timer = 2.0
+
+    if state is not None and weapon is not None:
+        from ..weapon_sfx import resolve_weapon_sfx_ref
+
+        key = resolve_weapon_sfx_ref(weapon.reload_sound)
+        if key is not None:
+            state.sfx_queue.append(key)
+
+
+def most_used_weapon_id_for_player(state: GameplayState, *, player_index: int, fallback_weapon_id: int) -> int:
+    """Return a 1-based weapon id for the player's most-used weapon."""
+
+    idx = int(player_index)
+    if 0 <= idx < len(state.weapon_shots_fired):
+        counts = state.weapon_shots_fired[idx]
+        if counts:
+            start = 1 if len(counts) > 1 else 0
+            best = max(range(start, len(counts)), key=lambda i: int(counts[i]))
+            if int(counts[best]) > 0:
+                return int(best)
+    return int(fallback_weapon_id)
+
+
+def player_swap_alt_weapon(player: PlayerState) -> bool:
+    """Swap primary and alternate weapon runtime blocks (Alternate Weapon perk)."""
+
+    if player.alt_weapon_id is None:
+        return False
+    (
+        player.weapon_id,
+        player.clip_size,
+        player.reload_active,
+        player.ammo,
+        player.reload_timer,
+        player.shot_cooldown,
+        player.reload_timer_max,
+        player.alt_weapon_id,
+        player.alt_clip_size,
+        player.alt_reload_active,
+        player.alt_ammo,
+        player.alt_reload_timer,
+        player.alt_shot_cooldown,
+        player.alt_reload_timer_max,
+    ) = (
+        player.alt_weapon_id,
+        player.alt_clip_size,
+        player.alt_reload_active,
+        player.alt_ammo,
+        player.alt_reload_timer,
+        player.alt_shot_cooldown,
+        player.alt_reload_timer_max,
+        player.weapon_id,
+        player.clip_size,
+        player.reload_active,
+        player.ammo,
+        player.reload_timer,
+        player.shot_cooldown,
+        player.reload_timer_max,
+    )
+    return True
+
+
+def player_start_reload(player: PlayerState, state: GameplayState) -> None:
+    """Start or refresh a reload timer (`player_start_reload` @ 0x00413430)."""
+
+    if player.reload_active and (
+        perk_active(player, PerkId.AMMUNITION_WITHIN) or perk_active(player, PerkId.REGRESSION_BULLETS)
+    ):
+        return
+
+    weapon = weapon_entry(player.weapon_id)
+    reload_time = float(weapon.reload_time) if weapon is not None and weapon.reload_time is not None else 0.0
+
+    if not player.reload_active:
+        player.reload_active = True
+
+    if perk_active(player, PerkId.FASTLOADER):
+        reload_time *= 0.7
+    if state.bonuses.weapon_power_up > 0.0:
+        reload_time *= 0.6
+
+    player.reload_timer = max(0.0, reload_time)
+    player.reload_timer_max = player.reload_timer
